+++ src/cmd/map.rs
@@ -1,7 +1,7 @@
 use std::convert::TryFrom;
 
 use crate::cmd::moonblade::{
-    run_moonblade_cmd, MoonbladeCmdArgs, LegacyMoonbladeErrorPolicy, MoonbladeMode,
+    run_moonblade_cmd, LegacyMoonbladeErrorPolicy, MoonbladeCmdArgs, MoonbladeMode,
 };
 use crate::config::Delimiter;
 use crate::util;
@@ -59,13 +59,9 @@ map options:
                                if you want the number of threads to be automatically chosen instead.
     -E, --errors <policy>      What to do with evaluation errors. One of:
                                  - "panic": exit on first error
-                                 - "report": add a column containing error
                                  - "ignore": coerce result for row to null
                                  - "log": print error to stderr
                                [default: panic].
-    --error-column <name>      Name of the column containing errors if -E/--errors
-                               is set to "report".
-                               [default: xan_error].
 
 Common options:
     -h, --help               Display this message
@@ -87,7 +83,6 @@ struct Args {
     flag_parallel: bool,
     flag_threads: Option<usize>,
     flag_errors: String,
-    flag_error_column: String,
 }
 
 pub fn run(argv: &[&str]) -> CliResult<()> {
@@ -108,7 +103,6 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         delimiter: args.flag_delimiter,
         parallelization,
         error_policy: LegacyMoonbladeErrorPolicy::try_from(args.flag_errors)?,
-        error_column_name: Some(args.flag_error_column),
         mode: MoonbladeMode::Map,
         ..Default::default()
     };