+++ tests/test_transform.rs
@@ -88,33 +88,6 @@ fn transform_errors_panic() {
 }
 
 #[test]
-fn transform_errors_report() {
-    let wrk = Workdir::new("transform_errors_report");
-    wrk.create(
-        "data.csv",
-        vec![svec!["a", "b"], svec!["1", "test"], svec!["2", "3"]],
-    );
-    let mut cmd = wrk.command("transform");
-    cmd.arg("b")
-        .arg("add(a, b)")
-        .args(&["-E", "report"])
-        .args(&["--error-column", "error"])
-        .arg("data.csv");
-
-    let got: Vec<Vec<String>> = wrk.read_stdout(&mut cmd);
-    let expected = vec![
-        svec!["a", "b", "error"],
-        svec![
-            "1",
-            "",
-            "add() cannot safely cast Bytes(\"test\") from type \"bytes\" to type \"number\""
-        ],
-        svec!["2", "5", ""],
-    ];
-    assert_eq!(got, expected);
-}
-
-#[test]
 fn transform_errors_ignore() {
     let wrk = Workdir::new("transform_errors_ignore");
     wrk.create(